requirements.txt
#Requirement for Duckerfile

fastapi
uvicorn
torch
torchvision
nest_asyncio
Pillow
